requirements.txt
mysql-connector-python==8.0.17
python-dotenv==1.0.0
selenium==4.13.0
chromedriver-binary==117.0.5938.149
